+++ CHANGELOG.md
@@ -6,9 +6,9 @@
 
 ## [Unreleased] - ReleaseDate
 
-### Fixed 
+### Added 
 
-- [#19](https://github.com/embedded-graphics/tinybmp/pull/19) Added support for color mapped 1bpp and 8bpp images.
+- [#19](https://github.com/embedded-graphics/tinybmp/pull/19) Added support for color mapped 1bpp and 8bpp images, along with `RawBmp::color_table` to get the raw color table data associated with the image.
 
 ## [0.3.1] - 2021-06-16