+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Added 
 
-- [#19](https://github.com/embedded-graphics/tinybmp/pull/19) Added support for color mapped 1bpp and 8bpp images, along with `RawBmp::color_table` to get the raw color table data associated with the image.
+- [#19](https://github.com/embedded-graphics/tinybmp/pull/19) Added support for color mapped 1bpp and 8bpp images. This change now also requires 1bpp and 8bpp images to contain a color table.
 
 ## [0.3.1] - 2021-06-16